requirements.txt
nltk
streamlit
speechrecognition
pyttsx3
python-docx
pandas
openpyxl